pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kairos-agent"
version = "0.1.0"
description = "Automated simulation content pipeline for short-form video production"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Orchestration
    "langgraph==0.3.30",
    "langchain-core>=0.3.0,<0.4",

    # LLM
    "litellm>=1.55.0,<2.0",
    "instructor>=1.7.0,<2.0",
    "anthropic>=0.42.0,<1.0",

    # Structured output / validation
    "pydantic>=2.10.0,<3.0",
    "pydantic-settings>=2.7.0,<3.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36,<3.0",
    "asyncpg>=0.30.0,<1.0",
    "psycopg2-binary>=2.9.10,<3.0",
    "alembic>=1.14.0,<2.0",

    # Task queue
    "redis>=5.2.0,<6.0",
    "celery>=5.4.0,<6.0",

    # RAG / Knowledge base
    "chromadb>=0.5.23,<1.0",
    "llama-index>=0.12.0,<1.0",

    # Web framework (review dashboard + API)
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.34.0,<1.0",
    "jinja2>=3.1.4,<4.0",
    "python-multipart>=0.0.18,<1.0",

    # Video / media
    "ffmpeg-python>=0.2.0,<1.0",
    "Pillow>=10.4.0,<11.0",

    # Observability
    "langfuse>=2.56.0,<3.0",

    # Utilities
    "httpx>=0.28.0,<1.0",
    "python-dotenv>=1.0.1,<2.0",
    "click>=8.1.7,<9.0",
    "rich>=13.9.0,<14.0",
    "tenacity>=9.0.0,<10.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0,<9.0",
    "pytest-asyncio>=0.25.0,<1.0",
    "pytest-cov>=6.0.0,<7.0",
    "pytest-timeout>=2.3.0,<3.0",
    "pytest-mock>=3.14.0,<4.0",

    # Linting & formatting
    "ruff>=0.9.0,<1.0",

    # Type checking
    "mypy>=1.14.0,<2.0",

    # Pre-commit
    "pre-commit>=4.0.0,<5.0",

    # Async test support
    "aiosqlite>=0.20.0,<1.0",
    "factory-boy>=3.3.0,<4.0",
]

[project.scripts]
pipeline = "kairos.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/kairos"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PT",     # flake8-pytest-style
    "PL",     # pylint
    "PERF",   # perflint
    "TRY",    # tryceratops
]
ignore = [
    "ANN101",  # Missing type annotation for self
    "ANN102",  # Missing type annotation for cls
    "ANN401",  # Dynamically typed expressions (typing.Any)
    "TRY003",  # Avoid specifying long messages outside the exception class
    "S603",    # subprocess call - check for untrusted input (we sandbox in Docker)
    "S607",    # Starting a process with a partial executable path
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Use of assert
    "ANN",     # Annotations not required in tests
    "ARG",     # Unused arguments common in fixtures
    "PLR2004", # Magic value comparisons OK in tests
]
"src/kairos/cli.py" = ["T20"]  # Print statements OK in CLI

[tool.ruff.lint.isort]
known-first-party = ["kairos"]

[tool.ruff.lint.pydantic]
# Ruff recognises Pydantic models for type-checking rules
root-are-public = true

# =============================================================================
# mypy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "chromadb.*",
    "ffmpeg.*",
    "langfuse.*",
    "litellm.*",
    "llama_index.*",
    "pymunk.*",
    "pygame.*",
    "redis.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

# =============================================================================
# pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Fast, no external dependencies",
    "integration: Needs DB, may need Docker",
    "llm: Calls real LLM (slow, costs money) — skip in CI",
    "quality: Runs on produced video files",
    "pipeline: Pipeline-specific tests (use with --pipeline=physics)",
    "slow: Tests taking >30s",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = "-v --tb=short"

# =============================================================================
# Coverage
# =============================================================================
[tool.coverage.run]
source = ["src/kairos"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "pass",
    "\\.\\.\\.",
]
